backend-services/fashion-clip-service/requirements.txt
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
filelock==3.15.4
flask==3.0.3
fsspec==2024.6.1
huggingface-hub==0.24.2
idna==3.7
importlib-metadata==8.2.0
itsdangerous==2.2.0
jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.13.1
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.2.2
tqdm==4.66.4
transformers==4.43.3
typing-extensions==4.12.2
urllib3==2.2.2
werkzeug==3.0.3
zipp==3.19.2
